Switch to char[] for data storage in-memory.
This change uses getBlob/setBlob to store byte[]-encoded char[]s in the
database, rather than getString/setString. It also tries its best to
destroy in-memory copies of things when fragments are stopped, etc.

HOWEVER, IT BREAKS SUBSTRING SEARCH FOR UNICODE SUBSTRINGS. That worked
before--I'm not sure why this breaks it, but it may be an encoding
incompatibility.